ts: move code to domains
