+++ typescript/src/shared/api.ts
@@ -1,21 +1,18 @@
 import Stripe from 'stripe';
+import {createCustomer, listCustomers} from './customers/functions';
+import {createProduct, listProducts} from './products/functions';
+import {createPrice, listPrices} from './prices/functions';
+import {createPaymentLink} from './paymentLinks/functions';
 import {
-  createCustomer,
-  listCustomers,
-  createProduct,
-  listProducts,
-  createPrice,
-  listPrices,
-  createPaymentLink,
   createInvoice,
   listInvoices,
   createInvoiceItem,
   finalizeInvoice,
-  retrieveBalance,
-  createRefund,
-  searchDocumentation,
-  listPaymentIntents,
-} from './functions';
+} from './invoices/functions';
+import {retrieveBalance} from './balance/functions';
+import {createRefund} from './refunds/functions';
+import {searchDocumentation} from './documentation/functions';
+import {listPaymentIntents} from './paymentIntents/functions';
 
 import type {Context} from './configuration';
