+++ typescript/src/shared/balance/functions.ts
@@ -1,0 +1,21 @@
+import Stripe from 'stripe';
+import {z} from 'zod';
+import {retrieveBalanceParameters} from './parameters';
+import type {Context} from '../configuration';
+
+export const retrieveBalance = async (
+  stripe: Stripe,
+  context: Context,
+  params: z.infer<ReturnType<typeof retrieveBalanceParameters>>
+) => {
+  try {
+    const balance = await stripe.balance.retrieve(
+      params,
+      context.account ? {stripeAccount: context.account} : undefined
+    );
+
+    return balance;
+  } catch (error) {
+    return 'Failed to retrieve balance';
+  }
+};
