+++ typescript/src/shared/balance/parameters.ts
@@ -1,0 +1,6 @@
+import {z} from 'zod';
+import type {Context} from '../configuration';
+
+export const retrieveBalanceParameters = (
+  _context: Context = {}
+): z.AnyZodObject => z.object({});
