+++ typescript/src/shared/balance/prompts.ts
@@ -1,0 +1,5 @@
+import type {Context} from '../configuration';
+
+export const retrieveBalancePrompt = (_context: Context = {}) => `
+This tool will retrieve the balance from Stripe. It takes no input.
+`;
