+++ typescript/src/shared/invoices/parameters.ts
@@ -1,0 +1,76 @@
+import {z} from 'zod';
+import type {Context} from '../configuration';
+
+export const createInvoiceParameters = (
+  context: Context = {}
+): z.AnyZodObject => {
+  const schema = z.object({
+    customer: z
+      .string()
+      .describe('The ID of the customer to create the invoice for.'),
+    days_until_due: z
+      .number()
+      .int()
+      .optional()
+      .describe('The number of days until the invoice is due.'),
+  });
+
+  if (context.customer) {
+    return schema.omit({customer: true});
+  } else {
+    return schema;
+  }
+};
+
+export const listInvoicesParameters = (
+  context: Context = {}
+): z.AnyZodObject => {
+  const schema = z.object({
+    customer: z
+      .string()
+      .optional()
+      .describe('The ID of the customer to list invoices for.'),
+    limit: z
+      .number()
+      .int()
+      .min(1)
+      .max(100)
+      .optional()
+      .describe(
+        'A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10.'
+      ),
+  });
+
+  if (context.customer) {
+    return schema.omit({customer: true});
+  } else {
+    return schema;
+  }
+};
+
+export const createInvoiceItemParameters = (
+  context: Context = {}
+): z.AnyZodObject => {
+  const schema = z.object({
+    customer: z
+      .string()
+      .describe('The ID of the customer to create the invoice item for.'),
+    price: z.string().describe('The ID of the price for the item.'),
+    invoice: z
+      .string()
+      .describe('The ID of the invoice to create the item for.'),
+  });
+
+  if (context.customer) {
+    return schema.omit({customer: true});
+  } else {
+    return schema;
+  }
+};
+
+export const finalizeInvoiceParameters = (
+  _context: Context = {}
+): z.AnyZodObject =>
+  z.object({
+    invoice: z.string().describe('The ID of the invoice to finalize.'),
+  });
