+++ typescript/src/shared/invoices/prompts.ts
@@ -1,0 +1,51 @@
+import type {Context} from '../configuration';
+
+export const createInvoicePrompt = (context: Context = {}) => {
+  const customerArg = context.customer
+    ? ''
+    : `- customer (str): The ID of the customer to create the invoice for.\n`;
+
+  return `
+This tool will create an invoice in Stripe.
+
+It takes ${context.customer ? 'one' : 'two'} argument${context.customer ? '' : 's'}:
+${customerArg}
+- days_until_due (int, optional): The number of days until the invoice is due.
+`;
+};
+
+export const listInvoicesPrompt = (context: Context = {}) => {
+  const customerArg = context.customer
+    ? ''
+    : `- customer (str, optional): The ID of the customer to list invoices for.\n`;
+
+  return `
+This tool will fetch a list of Invoices from Stripe.
+
+It takes ${context.customer ? 'one' : 'two'} argument${context.customer ? '' : 's'}:
+${customerArg}
+- limit (int, optional): The number of invoices to return.
+`;
+};
+
+export const createInvoiceItemPrompt = (context: Context = {}) => {
+  const customerArg = context.customer
+    ? ''
+    : `- customer (str): The ID of the customer to create the invoice item for.\n`;
+
+  return `
+This tool will create an invoice item in Stripe.
+
+It takes ${context.customer ? 'one' : 'two'} argument${context.customer ? '' : 's'}:
+${customerArg}
+- price (str): The ID of the price to create the invoice item for.
+- invoice (str): The ID of the invoice to create the invoice item for.
+`;
+};
+
+export const finalizeInvoicePrompt = (_context: Context = {}) => `
+This tool will finalize an invoice in Stripe.
+
+It takes one argument:
+- invoice (str): The ID of the invoice to finalize.
+`;
