+++ typescript/src/shared/paymentIntents/parameters.ts
@@ -1,0 +1,28 @@
+import {z} from 'zod';
+import type {Context} from '../configuration';
+
+export const listPaymentIntentsParameters = (
+  context: Context = {}
+): z.AnyZodObject => {
+  const schema = z.object({
+    customer: z
+      .string()
+      .optional()
+      .describe('The ID of the customer to list payment intents for.'),
+    limit: z
+      .number()
+      .int()
+      .min(1)
+      .max(100)
+      .optional()
+      .describe(
+        'A limit on the number of objects to be returned. Limit can range between 1 and 100.'
+      ),
+  });
+
+  if (context.customer) {
+    return schema.omit({customer: true});
+  } else {
+    return schema;
+  }
+};
