+++ typescript/src/shared/prices/prompts.ts
@@ -1,0 +1,18 @@
+import type {Context} from '../configuration';
+
+export const createPricePrompt = (_context: Context = {}) => `
+This tool will create a price in Stripe. If a product has not already been specified, a product should be created first.
+
+It takes three arguments:
+- product (str): The ID of the product to create the price for.
+- unit_amount (int): The unit amount of the price in cents.
+- currency (str): The currency of the price.
+`;
+
+export const listPricesPrompt = (_context: Context = {}) => `
+This tool will fetch a list of Prices from Stripe.
+
+It takes two arguments.
+- product (str, optional): The ID of the product to list prices for.
+- limit (int, optional): The number of prices to return.
+`;
