+++ typescript/src/shared/refunds/prompts.ts
@@ -1,0 +1,10 @@
+import type {Context} from '../configuration';
+
+export const createRefundPrompt = (_context: Context = {}) => `
+This tool will refund a payment intent in Stripe.
+
+It takes three arguments:
+- payment_intent (str): The ID of the payment intent to refund.
+- amount (int, optional): The amount to refund in cents.
+- reason (str, optional): The reason for the refund.
+`;
